# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gerenciamento de Consultas",
        "description": "Documentação da API para o projeto de agendamentos e serviços.",
        "contact": {
            "email": "seu.email@dominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desenvolvimento Local"
        }
    ],
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Lista todos os serviços (paginado)",
                "operationId": "listServices",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de serviços",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ServicesResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Cria um novo serviço",
                "operationId": "createService",
                "requestBody": {
                    "description": "Dados do serviço a ser criado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServicesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Serviço criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ServicesResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Busca um serviço pelo ID",
                "operationId": "getServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Serviço encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ServicesResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Atualiza um serviço existente",
                "operationId": "updateService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do serviço a ser atualizado",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateServicesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Serviço atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ServicesResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Exclui um serviço",
                "operationId": "deleteService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do serviço a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Serviço excluído com sucesso (Sem conteúdo)"
                    },
                    "404": {
                        "description": "Serviço não encontrado"
                    },
                    "422": {
                        "description": "Serviço em uso, não pode ser excluído"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreServicesRequest": {
                "title": "Store Service Request",
                "description": "Payload para criar um novo serviço",
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Equipamento"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 120.5
                    }
                },
                "type": "object"
            },
...
            "UpdateServicesRequest": {
                "title": "Update Service Request",
                "description": "Payload para atualizar um serviço",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Equipmento"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
            "ServicesResource": {
                "description": "Recurso que representa um Serviço",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Corte de cabelo"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token Sanctum (obrigatório para endpoints protegidos)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}